Treat hs2019 as rsa-sha256
This is stupid, but seems to be required for any kind of federation to
work with hs2019. hs2019 is *supposed* to mean the real algorithm is
sent OOB, but no fediverse implementations seem to do this. Instead,
they just assume it's a synonym for rsa-sha256. If we don't make this
assumption we can't validate hs2019 requests because they (mostly)
don't publish the real algorithm anywhere.